chore(typography): 🔧 refactored Title, Link, Text components
- attached element props right to styled component
- separated style file into own folder
