+++ src/components/Typography/Link/index.ts
@@ -1,1 +1,2 @@
+export * from './style';
 export * from './Link';
