+++ src/components/Typography/Link/style/index.ts
@@ -1,0 +1,1 @@
+export * from './Link.style';
