+++ src/components/Typography/Text/index.ts
@@ -1,2 +1,3 @@
-export * from './Text';
+export * from './style';
 export * from './types';
+export * from './Text';
