+++ src/components/Typography/Text/style/index.ts
@@ -1,0 +1,1 @@
+export * from './Text.style';
