+++ src/components/Typography/Title/index.ts
@@ -1,1 +1,2 @@
+export * from './style';
 export * from './Title';
