+++ src/components/Typography/Title/style/index.ts
@@ -1,0 +1,1 @@
+export * from './Title.style';
